# ddtrace/tracer.StartSpan: Use setMeta instead of SetTag