--- a/contrib/google.golang.org/grpc/grpc_test.go
+++ b/contrib/google.golang.org/grpc/grpc_test.go
@@ -799,7 +799,9 @@ func TestIgnoredMetadata(t *testing.T) {
 
 func BenchmarkUnaryServerInterceptor(b *testing.B) {
 	// need to use the real tracer to get representative measurments
-	tracer.Start(tracer.WithLogger(log.DiscardLogger{}))
+	tracer.Start(tracer.WithLogger(log.DiscardLogger{}),
+		tracer.WithEnv("test"),
+		tracer.WithServiceVersion("0.1.2"))
 	defer tracer.Stop()
 
 	doNothingOKGRPCHandler := func(ctx context.Context, req interface{}) (interface{}, error) {
@@ -868,6 +870,14 @@ func BenchmarkUnaryServerInterceptor(b *testing.B) {
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
 			interceptor(ctx, "ignoredRequestValue", methodInfo, doNothingErrorGRPCHandler)
+		}
+	})
+	interceptorNoStack := UnaryServerInterceptor(NoDebugStack())
+	b.Run("error_no_metadata_no_stack", func(b *testing.B) {
+		b.ReportAllocs()
+		b.ResetTimer()
+		for i := 0; i < b.N; i++ {
+			interceptorNoStack(ctx, "ignoredRequestValue", methodInfo, doNothingErrorGRPCHandler)
 		}
 	})
 }