--- a/ddtrace/tracer/tracer.go
+++ b/ddtrace/tracer/tracer.go
@@ -400,10 +400,10 @@ func (t *tracer) StartSpan(operationName string, options ...ddtrace.StartSpanOpt
 		delete(span.Metrics, keyMeasured)
 	}
 	if t.config.version != "" && span.Service == t.config.serviceName {
-		span.SetTag(ext.Version, t.config.version)
+		span.setMeta(ext.Version, t.config.version)
 	}
 	if t.config.env != "" {
-		span.SetTag(ext.Environment, t.config.env)
+		span.setMeta(ext.Environment, t.config.env)
 	}
 	if _, ok := span.context.samplingPriority(); !ok {
 		// if not already sampled or a brand new trace, sample it