@evanj 2022-02-07
We might want to merge PR #1135 first, because it adds a test that covers this code.