@felixge 2022-02-10
I've merged I merged https://github.com/DataDog/dd-trace-go/pull/1135  and updated this PR against master. So this should be ready to review for @DataDog/tracing-go .